refactor : code refactor
refactor : code refactor
1. Separate component for cardholder
2. Inline CSS removed
3. Proper line spacing.
4. change is Error msg
