+++ 06-team-data-github-api/src/component/CardHolder.js
@@ -1,0 +1,18 @@
+import CardComponent from './CardComponent';
+
+const CardHolder = ({ filteredTeam }) => {
+  if (filteredTeam.length > 0) {
+    return filteredTeam.map((user) => {
+      if (user.login) {
+        return (
+          <CardComponent
+            user={user}
+            key={user.id}
+          />
+        );
+      }
+    });
+  }
+};
+
+export default CardHolder;
