+++ 06-team-data-github-api/src/component/SearchComponent.js
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+
 const searchLogin = (searchKey, data) => {
   return data.filter((team) => {
     return (
