+++ 06-team-data-github-api/src/index.css
@@ -8,7 +8,9 @@
 .navbar {
   background-color: var(--primaryColor);
 }
-
+.card {
+  width: 16rem;
+}
 .cardFooter {
   background-color: var(--primaryColor);
   height: 3rem;
